pyproject.toml
[project]
name = "morning-glow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
]
